Split reviewer tests and tighten stale-thread smoke assertion
