+++ IntelligenceX.Tests/Program.Reviewer.CleanupAndThreads.cs
@@ -286,6 +286,7 @@ internal static partial class Program {
 
     private static void TestAutoResolveStaleThreadsSmoke() {
         var resolved = 0;
+        var resolvedThreadIdObserved = false;
         using var server = new LocalHttpServer(request => {
             if (!request.Path.Equals("/graphql", StringComparison.OrdinalIgnoreCase)) {
                 return null;
@@ -293,6 +294,7 @@ internal static partial class Program {
             if (!request.Body.Contains("resolveReviewThread", StringComparison.Ordinal)) {
                 return null;
             }
+            resolvedThreadIdObserved = request.Body.Contains("thread-old", StringComparison.Ordinal);
             resolved++;
             return new HttpResponse("{\"data\":{\"resolveReviewThread\":{\"thread\":{\"id\":\"thread-old\",\"isResolved\":true}}}}");
         });
@@ -313,6 +315,7 @@ internal static partial class Program {
 
         CallAutoResolveStaleThreads(github, threads, settings);
         AssertEqual(1, resolved, "stale auto-resolve smoke");
+        AssertEqual(true, resolvedThreadIdObserved, "stale auto-resolve targets thread-old");
     }
 
     private static void TestAutoResolveMissingInlineEmptyKeys() {
@@ -655,376 +658,6 @@ internal static partial class Program {
         var result = RunConfigValidation("{\"review\":{\"length\":\"SHORT\"}}");
         AssertEqual(true, result is not null, "validator result");
         AssertEqual(true, result!.Errors.Count > 0, "invalid enum should error");
-    }
-
-    private static void TestAnalysisSeverityCritical() {
-        AssertEqual("error", AnalysisSeverity.Normalize("critical"), "severity critical normalize");
-        AssertEqual(3, AnalysisSeverity.Rank("critical"), "severity critical rank");
-        AssertEqual("warning", AnalysisSeverity.Normalize("medium"), "severity medium normalize");
-    }
-
-    private static void TestAnalysisConfigExportToolIds() {
-        var temp = Path.Combine(Path.GetTempPath(), "ix-analysis-" + Guid.NewGuid().ToString("N"));
-        Directory.CreateDirectory(temp);
-        try {
-            var rules = new Dictionary<string, AnalysisRule>(StringComparer.OrdinalIgnoreCase) {
-                ["IX001"] = new AnalysisRule(
-                    "IX001", "csharp", "roslyn", "CA2000", "Dispose objects", "Ensure Dispose is called",
-                    "Reliability", "warning", Array.Empty<string>(), null, null),
-                ["IX002"] = new AnalysisRule(
-                    "IX002", "cs", "roslyn", "CA1062", "Validate arguments", "Validate argument null checks",
-                    "Reliability", "warning", Array.Empty<string>(), null, null),
-                ["PS001"] = new AnalysisRule(
-                    "PS001", "powershell", "psscriptanalyzer", "PSAvoidUsingWriteHost",
-                    "Avoid Write-Host", "Use Write-Output instead", "BestPractices", "suggestion",
-                    Array.Empty<string>(), null, null),
-                ["PS002"] = new AnalysisRule(
-                    "PS002", "ps", "psscriptanalyzer", "PSUseSupportsShouldProcess",
-                    "Use SupportsShouldProcess", "Add SupportsShouldProcess when needed", "BestPractices", "warning",
-                    Array.Empty<string>(), null, null)
-            };
-            var packs = new Dictionary<string, AnalysisPack>(StringComparer.OrdinalIgnoreCase) {
-                ["pack"] = new AnalysisPack(
-                    "pack", "Pack", "Test pack", new[] { "IX001", "IX002", "PS001", "PS002" },
-                    new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase), null)
-            };
-            var catalog = new AnalysisCatalog(rules, packs);
-            var settings = new AnalysisSettings { Packs = new[] { "pack" } };
-
-            AnalysisConfigExporter.Export(settings, catalog, temp);
-
-            var editorConfig = File.ReadAllText(Path.Combine(temp, ".editorconfig"));
-            AssertEqual(true, editorConfig.Contains("dotnet_diagnostic.CA2000.severity", StringComparison.Ordinal),
-                "editorconfig CA2000");
-            AssertEqual(true, editorConfig.Contains("dotnet_diagnostic.CA1062.severity", StringComparison.Ordinal),
-                "editorconfig CA1062");
-
-            var psConfig = File.ReadAllText(Path.Combine(temp, "PSScriptAnalyzerSettings.psd1"));
-            AssertEqual(true, psConfig.Contains("PSAvoidUsingWriteHost", StringComparison.Ordinal),
-                "psconfig PSAvoidUsingWriteHost");
-            AssertEqual(true, psConfig.Contains("PSAvoidUsingWriteHost = @{ Severity = 'Information' }", StringComparison.Ordinal),
-                "psconfig suggestion maps to information");
-            AssertEqual(true, psConfig.Contains("PSUseSupportsShouldProcess", StringComparison.Ordinal),
-                "psconfig PSUseSupportsShouldProcess");
-        } finally {
-            if (Directory.Exists(temp)) {
-                Directory.Delete(temp, true);
-            }
-        }
-    }
-
-    private static void TestAnalysisPolicyResolvesOverrides() {
-        var rules = new Dictionary<string, AnalysisRule>(StringComparer.OrdinalIgnoreCase) {
-            ["IX001"] = new AnalysisRule(
-                "IX001", "csharp", "roslyn", "CA2000", "Dispose objects", "Ensure Dispose is called",
-                "Reliability", "warning", Array.Empty<string>(), null, null),
-            ["IX002"] = new AnalysisRule(
-                "IX002", "powershell", "psscriptanalyzer", "PSAvoidUsingWriteHost", "Avoid Write-Host",
-                "Use Write-Output", "BestPractices", "warning", Array.Empty<string>(), null, null)
-        };
-        var packs = new Dictionary<string, AnalysisPack>(StringComparer.OrdinalIgnoreCase) {
-            ["default"] = new AnalysisPack(
-                "default", "Default", "Test pack", new[] { "IX001", "IX002" },
-                new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase) {
-                    ["IX001"] = "error"
-                }, null)
-        };
-        var catalog = new AnalysisCatalog(rules, packs);
-        var settings = new AnalysisSettings {
-            Packs = new[] { "default" },
-            SeverityOverrides = new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase) {
-                ["CA2000"] = "warning",
-                ["IX002"] = "error"
-            }
-        };
-
-        var policy = IntelligenceX.Analysis.AnalysisPolicyBuilder.Build(settings, catalog);
-        AssertEqual(0, policy.Warnings.Count, "policy warnings");
-        AssertEqual(2, policy.Rules.Count, "policy selected count");
-        AssertEqual("warning", policy.Rules["IX001"].Severity, "policy override tool rule id");
-        AssertEqual("error", policy.Rules["IX002"].Severity, "policy override catalog id");
-    }
-
-    private static void TestAnalysisPolicyResolvesIncludedPacks() {
-        var rules = new Dictionary<string, AnalysisRule>(StringComparer.OrdinalIgnoreCase) {
-            ["IX001"] = new AnalysisRule(
-                "IX001", "csharp", "roslyn", "CA2000", "Dispose objects", "Ensure Dispose is called",
-                "Reliability", "warning", Array.Empty<string>(), null, null),
-            ["IX002"] = new AnalysisRule(
-                "IX002", "csharp", "roslyn", "CA1062", "Validate arguments", "Validate public arguments",
-                "Reliability", "warning", Array.Empty<string>(), null, null),
-            ["IX003"] = new AnalysisRule(
-                "IX003", "csharp", "roslyn", "SA1600", "Document elements", "Document public symbols",
-                "Style", "warning", Array.Empty<string>(), null, null)
-        };
-        var packs = new Dictionary<string, AnalysisPack>(StringComparer.OrdinalIgnoreCase) {
-            ["core"] = new AnalysisPack(
-                "core", "Core", "Core rules", new[] { "IX001" },
-                new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase), null),
-            ["standard"] = new AnalysisPack(
-                "standard", "Standard", "Standard rules", new[] { "IX002" },
-                new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase) {
-                    ["IX001"] = "error"
-                }, null, includes: new[] { "core" }),
-            ["strict"] = new AnalysisPack(
-                "strict", "Strict", "Strict rules", new[] { "IX003" },
-                new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase) {
-                    ["IX002"] = "error"
-                }, null, includes: new[] { "standard" })
-        };
-        var catalog = new AnalysisCatalog(rules, packs);
-        var settings = new AnalysisSettings {
-            Packs = new[] { "strict" }
-        };
-
-        var policy = IntelligenceX.Analysis.AnalysisPolicyBuilder.Build(settings, catalog);
-        AssertEqual(0, policy.Warnings.Count, "policy include warnings");
-        AssertEqual(3, policy.Rules.Count, "policy include selected count");
-        AssertEqual("error", policy.Rules["IX001"].Severity, "policy include inherited override");
-        AssertEqual("error", policy.Rules["IX002"].Severity, "policy include parent override");
-        AssertEqual("warning", policy.Rules["IX003"].Severity, "policy include direct rule");
-    }
-
-    private static void TestAnalysisPolicyIncludedPackCycleWarning() {
-        var rules = new Dictionary<string, AnalysisRule>(StringComparer.OrdinalIgnoreCase) {
-            ["IX001"] = new AnalysisRule(
-                "IX001", "csharp", "roslyn", "CA2000", "Dispose objects", "Ensure Dispose is called",
-                "Reliability", "warning", Array.Empty<string>(), null, null),
-            ["IX002"] = new AnalysisRule(
-                "IX002", "csharp", "roslyn", "CA1062", "Validate arguments", "Validate public arguments",
-                "Reliability", "warning", Array.Empty<string>(), null, null)
-        };
-        var packs = new Dictionary<string, AnalysisPack>(StringComparer.OrdinalIgnoreCase) {
-            ["pack-a"] = new AnalysisPack(
-                "pack-a", "Pack A", "A rules", new[] { "IX001" },
-                new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase), null, includes: new[] { "pack-b" }),
-            ["pack-b"] = new AnalysisPack(
-                "pack-b", "Pack B", "B rules", new[] { "IX002" },
-                new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase), null, includes: new[] { "pack-a" })
-        };
-        var catalog = new AnalysisCatalog(rules, packs);
-        var settings = new AnalysisSettings {
-            Packs = new[] { "pack-a" }
-        };
-
-        var policy = IntelligenceX.Analysis.AnalysisPolicyBuilder.Build(settings, catalog);
-        AssertEqual(true, policy.Rules.ContainsKey("IX001"), "policy cycle includes pack-a rules");
-        AssertEqual(true, policy.Rules.ContainsKey("IX002"), "policy cycle includes pack-b rules");
-        AssertEqual(true, policy.Warnings.Any(warning => warning.Contains("Pack include cycle detected", StringComparison.OrdinalIgnoreCase)),
-            "policy cycle warning");
-    }
-
-    private static void TestAnalysisCatalogValidatorPassesBuiltInCatalog() {
-        var workspace = ResolveWorkspaceRoot();
-        var validation = IntelligenceX.Analysis.AnalysisCatalogValidator.ValidateWorkspace(workspace);
-        AssertEqual(true, validation.IsValid, "catalog validator built-in valid");
-        AssertEqual(0, validation.Errors.Count, "catalog validator built-in errors");
-    }
-
-    private static void TestAnalysisCatalogValidatorDetectsInvalidCatalog() {
-        var temp = Path.Combine(Path.GetTempPath(), "ix-analysis-validate-" + Guid.NewGuid().ToString("N"));
-        Directory.CreateDirectory(temp);
-        try {
-            var rulesDir = Path.Combine(temp, "Analysis", "Catalog", "rules", "csharp");
-            var packsDir = Path.Combine(temp, "Analysis", "Packs");
-            Directory.CreateDirectory(rulesDir);
-            Directory.CreateDirectory(packsDir);
-
-            File.WriteAllText(Path.Combine(rulesDir, "IX001.json"), """
-{
-  "id": "IX001",
-  "language": "csharp",
-  "tool": "roslyn",
-  "title": "Rule one",
-  "description": "Rule one"
-}
-""");
-            File.WriteAllText(Path.Combine(rulesDir, "IX001.duplicate.json"), """
-{
-  "id": "IX001",
-  "language": "csharp",
-  "tool": "roslyn",
-  "title": "Rule one duplicate",
-  "description": "Rule one duplicate"
-}
-""");
-            File.WriteAllText(Path.Combine(packsDir, "pack-a.json"), """
-{
-  "id": "pack-a",
-  "label": "Pack A",
-  "includes": ["pack-b", "pack-missing"],
-  "rules": ["IX001", "IX404"],
-  "severityOverrides": {
-    "IX001": "banana"
-  }
-}
-""");
-            File.WriteAllText(Path.Combine(packsDir, "pack-b.json"), """
-{
-  "id": "pack-b",
-  "label": "Pack B",
-  "includes": ["pack-a"],
-  "rules": []
-}
-""");
-
-            var validation = IntelligenceX.Analysis.AnalysisCatalogValidator.ValidateWorkspace(temp);
-            AssertEqual(false, validation.IsValid, "catalog validator invalid catalog");
-            AssertEqual(true, validation.Errors.Any(error => error.Contains("Duplicate rule id 'IX001'", StringComparison.OrdinalIgnoreCase)),
-                "catalog validator duplicate rule id");
-            AssertEqual(true, validation.Errors.Any(error => error.Contains("unknown rule 'IX404'", StringComparison.OrdinalIgnoreCase)),
-                "catalog validator unknown rule");
-            AssertEqual(true, validation.Errors.Any(error => error.Contains("unknown pack 'pack-missing'", StringComparison.OrdinalIgnoreCase)),
-                "catalog validator unknown include");
-            AssertEqual(true, validation.Errors.Any(error => error.Contains("unsupported severity 'banana'", StringComparison.OrdinalIgnoreCase)),
-                "catalog validator unsupported severity");
-            AssertEqual(true, validation.Errors.Any(error => error.Contains("Pack include cycle detected", StringComparison.OrdinalIgnoreCase)),
-                "catalog validator include cycle");
-        } finally {
-            if (Directory.Exists(temp)) {
-                Directory.Delete(temp, true);
-            }
-        }
-    }
-
-    private static void TestAnalysisCatalogValidatorDetectsMissingRuleMetadata() {
-        var temp = Path.Combine(Path.GetTempPath(), "ix-analysis-validate-missing-field-" + Guid.NewGuid().ToString("N"));
-        Directory.CreateDirectory(temp);
-        try {
-            var rulesDir = Path.Combine(temp, "Analysis", "Catalog", "rules", "csharp");
-            var packsDir = Path.Combine(temp, "Analysis", "Packs");
-            Directory.CreateDirectory(rulesDir);
-            Directory.CreateDirectory(packsDir);
-
-            File.WriteAllText(Path.Combine(rulesDir, "IXMISSING.json"), """
-{
-  "id": "IXMISSING",
-  "language": "csharp",
-  "tool": "roslyn",
-  "description": "Missing title"
-}
-""");
-
-            var validation = IntelligenceX.Analysis.AnalysisCatalogValidator.ValidateWorkspace(temp);
-            AssertEqual(false, validation.IsValid, "catalog validator missing required metadata invalid");
-            AssertEqual(true,
-                validation.Errors.Any(error => error.Contains("Rule 'IXMISSING' missing required field 'title'",
-                    StringComparison.OrdinalIgnoreCase)),
-                "catalog validator missing required metadata");
-        } finally {
-            if (Directory.Exists(temp)) {
-                Directory.Delete(temp, true);
-            }
-        }
-    }
-
-    private static void TestAnalyzeValidateCatalogCommand() {
-        var temp = Path.Combine(Path.GetTempPath(), "ix-analyze-validate-command-" + Guid.NewGuid().ToString("N"));
-        Directory.CreateDirectory(temp);
-        try {
-            var rulesDir = Path.Combine(temp, "Analysis", "Catalog", "rules", "csharp");
-            var packsDir = Path.Combine(temp, "Analysis", "Packs");
-            Directory.CreateDirectory(rulesDir);
-            Directory.CreateDirectory(packsDir);
-
-            File.WriteAllText(Path.Combine(rulesDir, "IX001.json"), """
-{
-  "id": "IX001",
-  "language": "csharp",
-  "tool": "roslyn",
-  "title": "Rule one",
-  "description": "Rule one"
-}
-""");
-            File.WriteAllText(Path.Combine(packsDir, "pack-a.json"), """
-{
-  "id": "pack-a",
-  "label": "Pack A",
-  "rules": ["IX001"]
-}
-""");
-
-            var validExit = IntelligenceX.Cli.Analysis.AnalyzeRunner.RunAsync(new[] {
-                "validate-catalog",
-                "--workspace",
-                temp
-            }).GetAwaiter().GetResult();
-            AssertEqual(0, validExit, "analyze validate-catalog valid exit");
-
-            File.WriteAllText(Path.Combine(packsDir, "pack-a.json"), """
-{
-  "id": "pack-a",
-  "label": "Pack A",
-  "rules": ["IX404"]
-}
-""");
-
-            var invalidExit = IntelligenceX.Cli.Analysis.AnalyzeRunner.RunAsync(new[] {
-                "validate-catalog",
-                "--workspace",
-                temp
-            }).GetAwaiter().GetResult();
-            AssertEqual(1, invalidExit, "analyze validate-catalog invalid exit");
-        } finally {
-            if (Directory.Exists(temp)) {
-                Directory.Delete(temp, true);
-            }
-        }
-    }
-
-    private static void TestAnalysisCatalogRuleDocsPath() {
-        var temp = Path.Combine(Path.GetTempPath(), "ix-analysis-docs-" + Guid.NewGuid().ToString("N"));
-        Directory.CreateDirectory(temp);
-        try {
-            var rulesDir = Path.Combine(temp, "Analysis", "Catalog", "rules", "internal");
-            var packsDir = Path.Combine(temp, "Analysis", "Packs");
-            Directory.CreateDirectory(rulesDir);
-            Directory.CreateDirectory(packsDir);
-            Directory.CreateDirectory(Path.Combine(temp, "Docs", "reviewer"));
-
-            var docsPath = "Docs/reviewer/static-analysis.md";
-            File.WriteAllText(Path.Combine(temp, docsPath), "# docs");
-            File.WriteAllText(Path.Combine(rulesDir, "IXLOC001.json"), """
-{
-  "id": "IXLOC001",
-  "language": "internal",
-  "tool": "IntelligenceX.Maintainability",
-  "toolRuleId": "IXLOC001",
-  "title": "Source files should stay below 700 lines",
-  "description": "Flags oversized source files.",
-  "category": "Maintainability",
-  "defaultSeverity": "warning",
-  "docs": "Docs/reviewer/static-analysis.md"
-}
-""");
-
-            var catalog = IntelligenceX.Analysis.AnalysisCatalogLoader.LoadFromWorkspace(temp);
-            AssertEqual(true, catalog.Rules.TryGetValue("IXLOC001", out var rule), "analysis docs rule exists");
-            AssertEqual(docsPath, rule!.Docs, "analysis docs path stored");
-            var resolvedDocsPath = Path.Combine(temp, rule.Docs!.Replace('/', Path.DirectorySeparatorChar));
-            AssertEqual(true, File.Exists(resolvedDocsPath), "analysis docs path resolves from workspace");
-        } finally {
-            if (Directory.Exists(temp)) {
-                Directory.Delete(temp, true);
-            }
-        }
-    }
-
-    private static string ResolveWorkspaceRoot() {
-        var current = Environment.CurrentDirectory;
-        for (var i = 0; i < 12; i++) {
-            var rulesDir = Path.Combine(current, "Analysis", "Catalog", "rules");
-            var packsDir = Path.Combine(current, "Analysis", "Packs");
-            if (Directory.Exists(rulesDir) && Directory.Exists(packsDir)) {
-                return current;
-            }
-            var parent = Directory.GetParent(current);
-            if (parent is null) {
-                break;
-            }
-            current = parent.FullName;
-        }
-        return Environment.CurrentDirectory;
     }
 
 }
